style(http): Add Seperate handlers for http and ws
